Add support for stdout output

+++ isort/api.py
@@ -422,38 +422,50 @@ def sort_file(
     disregard_skip: bool = False,
     ask_to_apply: bool = False,
     show_diff: bool = False,
+    output_to_stdout: bool = False
     **config_kwargs,
 ):
     with io.read_file(filename) as source_file:
-        tmp_file = source_file.path.with_suffix(source_file.path.suffix + ".isorted")
         changed: bool = False
         try:
-            with tmp_file.open("w", encoding=source_file.encoding, newline="") as output_stream:
+            if output_to_stdout:
                 changed = sorted_imports(
                     input_stream=source_file.stream,
-                    output_stream=output_stream,
+                    output_stream=sys.stdout,
                     config=config,
                     file_path=source_file.path,
                     disregard_skip=disregard_skip,
                     **config_kwargs
                 )
-            if changed:
-                if show_diff or ask_to_apply:
-                    source_file.stream.seek(0)
-                    show_unified_diff(
-                        file_input=source_file.stream.read(),
-                        file_output=tmp_file.read_text(encoding=source_file.encoding),
-                        file_path=source_file.path
-                    )
-                    if ask_to_apply and not ask_whether_to_apply_changes_to_file(str(source_file.path)):
-                        return
-                tmp_file.replace(source_file.path)
+            else:
+                tmp_file = source_file.path.with_suffix(source_file.path.suffix + ".isorted")
+                try:
+                    with tmp_file.open("w", encoding=source_file.encoding, newline="") as output_stream:
+                        changed = sorted_imports(
+                            input_stream=source_file.stream,
+                            output_stream=output_stream,
+                            config=config,
+                            file_path=source_file.path,
+                            disregard_skip=disregard_skip,
+                            **config_kwargs
+                        )
+                    if changed:
+                        if show_diff or ask_to_apply:
+                            source_file.stream.seek(0)
+                            show_unified_diff(
+                                file_input=source_file.stream.read(),
+                                file_output=tmp_file.read_text(encoding=source_file.encoding),
+                                file_path=source_file.path
+                            )
+                            if ask_to_apply and not ask_whether_to_apply_changes_to_file(str(source_file.path)):
+                                return
+                        tmp_file.replace(source_file.path)
+                finally:
+                    try:  # Python 3.8+: use `missing_ok=True` instead of try except.
+                        tmp_file.unlink()
+                    except FileNotFoundError:
+                        pass
         except ExistingSyntaxErrors:
             warn("{file_path} unable to sort due to existing syntax errors")
         except IntroducedSyntaxErrors:
             warn("{file_path} unable to sort as isort introduces new syntax errors")
-        finally:
-            try:  # Python 3.8+: use `missing_ok=True` instead of try except.
-                tmp_file.unlink()
-            except FileNotFoundError:
-                pass
